// web/src/api/generated/models/preparationBulkAdvanceOutcome.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { PreparationUnitResponse } from './preparationUnitResponse';

export interface PreparationBulkAdvanceOutcome {
  code?: string;
  preparation_unit_id?: string;
  status?: string;
  unit?: PreparationUnitResponse;
}